requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
pydantic
faiss-cpu
python-dotenv
Langchain-groq
